qt-designer/dataviewer/requirements.txt
PyQt5
PyInstaller
pycryptodome
